refactor: WatchListItem 제거 및 WatchList 변경
