+++ backend/src/main/java/kakaobootcamp/backend/domains/member/domain/Member.java
@@ -14,6 +14,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import kakaobootcamp.backend.domains.transaction.domain.Transaction;
+import kakaobootcamp.backend.domains.watchList.domain.WatchList;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -69,6 +70,9 @@ public class Member extends BaseEntity {
 
 	@OneToMany(mappedBy = "member", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
 	private List<Transaction> transactions = new ArrayList<>();
+
+	@OneToMany(mappedBy = "member", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+	private List<WatchList> watchLists = new ArrayList<>();
 
 	@Builder
 	private Member(
